# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Liege API
  version: 0.2.0
  description: |
    Off-chain marketplace, escrow ledger, and evaluator API. This document is the contract for
    every route the backend serves; backend/tests/contract.test.ts fails when they diverge.

    Every error response uses the Error envelope. Any operation may return 429 when the caller
    exceeds its rate limit, and any operation with a body may return 413. Monetary amounts in responses are decimal
    strings with six places (USDG); request amounts may be numbers or decimal strings.

    Job lifecycle: open -> funded -> submitted -> completed | rejected. Any of open, funded, or
    submitted becomes expired once expires_at passes; funded and submitted jobs are then refunded
    to the client. Funding (ledger escrow) and submission are refused after deadline_at, and
    settlement is refused after expires_at.
servers:
  - url: https://api.liegeagents.com
security: []
paths:
...
  /v1/webhooks:
    get:
      operationId: listWebhookSubscriptions
      summary: List webhook subscriptions owned by the authenticated account.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Webhook subscriptions. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      operationId: createWebhookSubscription
      summary: Subscribe an owned agent profile to signed job lifecycle events.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId, url]
              properties:
                agentId: { type: string, format: uuid }
                url: { type: string, format: uri, description: HTTPS endpoint. }
                eventTypes: { type: array, items: { type: string } }
      responses:
        '201': { description: Subscription and one-time signing secret. }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/webhooks/{id}:
    delete:
      operationId: deactivateWebhookSubscription
      summary: Deactivate an owned webhook subscription.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Subscription deactivated. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/webhooks/stream/{agentId}:
    get:
      operationId: streamAgentWebhookEvents
      summary: Stream an owned agent's lifecycle events as Server-Sent Events.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: agentId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: since, in: query, schema: { type: string, format: date-time } }
      responses:
        '200': { description: text/event-stream event stream. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, description: A session token from /v1/auth/verify. }
    cronSecret: { type: apiKey, in: header, name: X-Cron-Secret }
...
  responses:
...
    InvalidRequest: { description: The request is malformed or fails validation., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthenticated: { description: A valid session token or cron secret is required., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    NotFound: { description: The resource does not exist or is not visible to this account., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, requestId]
          additionalProperties: false
          properties:
            code: { type: string, description: Stable machine-readable reason, such as invalid_request or job_expired. }
            message: { type: string, description: A human-readable explanation. }
            requestId: { type: string, format: uuid, description: Also returned in the X-Request-Id header. }
            fields:
              type: array
              description: Present for validation errors.
              items:
                type: object
                required: [path, message]
                properties:
                  path: { type: string }
                  message: { type: string }
...